+++ Taxamo/Model/InputTransactionLine.cs
@@ -23,7 +23,7 @@ namespace Taxamo.Model {
     
     /* Unit price. */
     [DataMember(Name="unit_price", EmitDefaultValue=false)]
-    public double? UnitPrice { get; set; }
+    public decimal? UnitPrice { get; set; }
 
     
     /* Unit of measure. */
@@ -33,7 +33,7 @@ namespace Taxamo.Model {
     
     /* Quantity Defaults to 1. */
     [DataMember(Name="quantity", EmitDefaultValue=false)]
-    public double? Quantity { get; set; }
+    public decimal? Quantity { get; set; }
 
     
     /* Custom fields, stored as key-value pairs. This property is not processed and used mostly with Taxamo-built helpers. */
@@ -58,7 +58,7 @@ namespace Taxamo.Model {
     
     /* Amount. Required if total amount is not provided. */
     [DataMember(Name="amount", EmitDefaultValue=false)]
-    public double? Amount { get; set; }
+    public decimal? Amount { get; set; }
 
     
     /* Custom id, provided by ecommerce software. */
@@ -73,12 +73,12 @@ namespace Taxamo.Model {
     
     /* Tax rate, calculated by taxamo. Must be provided when informative field is true. */
     [DataMember(Name="tax_rate", EmitDefaultValue=false)]
-    public double? TaxRate { get; set; }
+    public decimal? TaxRate { get; set; }
 
     
     /* Total amount. Required if amount is not provided. */
     [DataMember(Name="total_amount", EmitDefaultValue=false)]
-    public double? TotalAmount { get; set; }
+    public decimal? TotalAmount { get; set; }
 
     
     /* Line contents description. */
